+++ server/app.js
@@ -1,1 +1,4 @@
-// main express file
+const express = require('express');
+const app = express();
+
+module.exports = app;
