+++ server/index.js
@@ -1,1 +1,11 @@
-// port and listening file, also calls seed file 
+const app = require('./app')
+const syncAndSeed = require('./server/db/seed');
+const PORT = process.env.PORT || 3000;
+
+const init = () => {
+  return syncAndSeed();
+};
+
+app.listen(PORT, () => console.log(`listening on port ${PORT}`));
+
+init();
